Add files via upload
This work was an analysis of different variables of distinct kinds to examine the relationship between categorical and demographic variables with the results of the admission exam of the National University of Colombia. So the principal main goal to us was to predict the behavior of data and looking for the best model to do that. You can find a big quantity of models, in distinct approaches but with the same purpose.

PD: Angie :3