Add files via upload
This document is the analysis conducted on the relationship between alcohol and cigarette consumption with various demographic variables. This analysis is carried out without prejudice and is done purely out of curiosity. In it, you will find the final conclusion and the different phases of creating the model, with the aim of identifying the truly important variables for our study, as well as determining if the model makes sense through various statistical significance tests.

The main question we wanted to answer was: Is early consumption of these psychoactive substances a determining factor in generating addiction in a person?

Enjoy this work.